03 metadata_analysis/requirements.txt
pillow==11.0.0
python-magic==0.4.27
